+++ NEWS.md
@@ -2,6 +2,7 @@
 ## Minor changes
 * Check valid columns in `ExtractGEOMeta`.
 * Add `timeout` to `ExtractGEOExpSupp` , `DownloadGEOSupp` and `ParseGEO`.
+* Add `tryCatch` in `ParseGEO` and `DownloadGEOSupp`.
 
 
 -------------